requirements.txt
anyio==4.6.2.post1
beautifulsoup4==4.12.3
certifi==2024.8.30
click==8.1.7
fastcore==1.7.20
fastlite==0.0.13
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
oauthlib==3.2.2
packaging==24.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-fasthtml==0.10.0
python-multipart==0.0.17
PyYAML==6.0.2
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlite-minutils==4.0.2
starlette==0.41.3
uv==0.5.4
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
